+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* `flexberry-lookup` component:
+    * Fix `placeholder` in dropdown mode.
 
 ## [0.8.4] - 2017-02-07
 ### Changed